+++ pyqmix/config.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+DLL_DIR = None
+CONFIG_DIR = None